+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-[Unreleased]: https://github.com/nik-rev/cargo-reedme/compare/v0.3.6...HEAD
+[Unreleased]: https://github.com/nik-rev/cargo-reedme/compare/v0.4.0...HEAD
+
+## [v0.4.0] - 2026-02-27
+
+[v0.4.0]: https://github.com/nik-rev/cargo-reedme/compare/v0.3.6...v0.4.0
 
 ### Changed